fix: 컨테이너 분리 + 삼항연산자 사용
